Updated bot.py so that it works for me
I will get  the error that NICK, PASS, and ROOM are not defined if I try to start the bot without having settings.txt. This can be avoided by having empty NICK, PASS, and ROOM variables (which will get changed to the normal values if settings.txt is existent.
I changed the text mode in line 56 from w to w+, which means that settings.txt will automatically get created if it doesn't exist. Apparently, this happens even if the textmode is w ... but it shouldn't.